Defects: 272, 259, 260 (interface cli commands)
259: There should be a CLI command to configure split modules
260: CLI should display pluggable module information
272: Turning off Auto-negotiation on QSFP ports changes admin state

Implement commands to turn on and off split modules.
Extend "show interface" commands to allow display of pluggable module
info.
Display "error" column when not empty to indicate why the interface
is adminstratively down. In the case of 272, auto-negotiation is
required for QSFP_CR4 modules, and this misconfiguration was detected
and reported in the "error" column.

Includes an example of sorting.

Change-Id: Ida234e1337d09d6cd65a435c71aa44e240005197
Signed-off-by: Dan Cripe <dan.cripe@hpe.com>