Code diff for user story # 2793
Problem:-
sh rib not showing unselected routes and next hops - CLI bug

Solution:-
In the show_rib() function in the file src/ops-cli/vtysh/l3routes_vty.o for the routers which are marked slected 'false', we are not displaying the next-hops while walking the next-hops in the "show rib" command. We need to remove that check so that all the next-hops of an unselected route also get displayed.

Change-Id: I56610ce0aec5cdfcf1cbdacb0cab644aa27e4d13
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>