Fix for Issue: 82 in CLI
This commit includes changes to prevent user from modifying
or deleting an internal VLAN. To maintain consistency,
vlan names written to DB will use 'VLAN' instead of 'vlan'
as internal vlans use 'VLAN' and this is the convention used
in ovs-vsctl commands.
Modified scripts to work with the new name change and added
a test case for internal vlan checks.

Change-Id: Ie1cac1569c80a57cf2684374e888bb4573019693
Signed-off-by: Vasanth Viswanathan <vasanth.viswanathan@hpe.com>