Fixed Critical Defect# Taiga:100 LACP intf status not visible from VTYSH
Modified CLI commands:
  show lacp interfaces <cr> && show lacp interfaces IFNAME <cr>
Changes:
Added the lacp_status column to the IDL cache in order to read from it
Added the System-id and System-priority to the output
It now displays the same list of columns in local and partner views
Modified the output to show the columns in same order for consistency

Before this change the priority is obtained from the other_config table but
this table is only loaded after the user explicitly configures it.
However during normal operation all interfaces have a default priority.
Also, because the priority and port are configured separately, it makes
sense to present them in two dift columns.
For consistency purposes, the same list of columns are being displayed in
both the local and partner outputs.
As per requirement, the output should display the IDs and priorities.

Change-Id: Ic9dfedcdd19905ea6ee62bdb1838cd1ec00c28b0
Signed-off-by: Walinton Cambronero <walinton.cambronero@hpe.com>