User Story ID-3081: Update show interface command
Add new functionality to "show interface" command so it shows lacp aggregates configured and its statistics.
The value of the statistics are the sum of the statistics of each interface aggregated to the LAG.
Now, It is possible to execute "show interface [interface/lag_name] [brief]" to show specific information about LAGs.

Change-Id: I7eae85abd7b0b04e9a0c5d3269463b23310796ab
Signed-off-by: Agustin Meneses <agustin-jose.meneses-fuentes@hpe.com>