Fixed Defect #358: Vlan lag configuration don’t overwrite interface vlan configuration
Changes:
Fixed the way the command was removing a port entry of interface if already exist,
added a function to remove the port reference from bridge or VRF first.
Before this change, if the port already exist, the interface will be duplicated in
different ports, one in the original and the other in the LAG port.
With this change the original port will be deleted and the interface moved to the LAG port.

Change-Id: I7ae53107b99086dab3e45f4d64ed29cf1b61f231
Signed-off-by: David Herdoiza Morales <david.herdoiza@hpe.com>