fix dev missed NULL check that breaks vtysh debug
This bug prevents vtysh from running when built with debug flags
(it segfaults on the modified line).

A line in the next for() loop looks exactly like this (fixed) line,
indicating perhaps a copy/paste omission or testing gap.

It's unclear from comments and context why this function even exists
in the first place, but that's beyond the scope of this fix.

Fix has been verified on my local system, as it was required in order to
do some debug on my own code so I could even run debug-enabled vtysh.

Change-Id: Iad65ab20c4fe4a7246991557bcbe4582c10af6da
Signed-off-by: Kevin Worth <kevin.worth@hpe.com>