fix: dev: Bug #829: Memory corruption when ports are attached to a named VRF
With default VRF having one port, and trying to move the port to a
named VRF, xmalloc of zero is being done, which is returning an address
and this address is later given to xrealloc, which is failing to realloc
to the new size and is giving the same previous memory. The memory is of
size zero.In such scenario, this issue is hit.

To avoid this scenario as a fix - freeing the previously provided memory and
xmallocing to get the memory for ports and use it furthur.

Change-Id: Id8505d08d7f4c7ea0a2af83d9470ea601b0bec53
Signed-off-by: Jayachandra Babu K <jayachandra.bab.k@hpe.com>