Fix for Bug: 854
system_vtysh_utils.h was defining the output values of the
status of power supply. But it was missing extern keyword
and the .c file was never included as part of the build.
This was causing powerd to crash and sysd to display null values.

Included a simple fix to define all const values.

Change-Id: I8cba6e98c1d2c5151076c47ba542c821d8ffe3bc
Signed-off-by: Jui Shah <jui.d.shah@hpe.com>