Crash with 500 routes with interface
In situations when a port is moved from being a routed port to 
a switch port and vice versa, it might happen that the 'show rib' 
might refer to an deallocated port. This movement of the port is
a delete of the old port and creation of a new port with the relevant 
router/switch type detail which is mapped to Bridge/Route table 
accordingly. So to prevent a crash when such a movement happens 
NULL checks have to be in place before printing/writing to the port
name.

TG-760
Tags: fix, dev

Change-Id: Ie4a3269d38fa9988999b47f2e900657a86a9b13d
Signed-off-by: Srinivasan Srivatsan <srinivasan.srivatsan@hpe.com>