fix: dev: Fix bug TG-1191 ALM#2507: Physical interface not in sort order on “show vrf’ output
This fix includes sorting of the Interfaces in the display command for VRF.
As a part of fix , in qsort compare function, checking if both the elements are digits
and returning 0,1 or -1 based on the value comparision. Else if one of the elements
is a string then using strcmp to return the value.

Change-Id: I63af57077e8abe6c8bea6c1cbc5184d2c5aeedd8
Signed-off-by: Jayachandra Babu K <jayachandra.bab.k@hpe.com>