Re-skip a ct that should not have been un-skipped
This test case:
test_vtysh_ct_proxy_arp

was being skipped on master, and is still being skipped on rel/dill. Review
9176 un-skipped the test:
Commit 8ff7d5a7b09fb673b7297a17820c8871e02220e6

This results in the current state of ops-cli, where the test case fails in a
fresh clone of master, preventing any merges to ops-cli.

This commit re-skips the test that should not have been un-skipped, thus
putting ops-cli back into a state where changes can be merged.

Once the ct works, then it can be re-enabled.

Change-Id: Ia5c7695dd6148ae0738b6770893bf68efcc7bb82
Signed-off-by: Miles Cote <miles.cote@hpe.com>