fix, usr: Modified Redistribute CLI Support
TG - 1108
Severity - High

Issue :

The current schema implementation for bgp redistribute configuration is
a key-value pair where the key is the protocol and value is UUID of the
Route Map table entry. With this implementation,OVSDB always expect UUID
as the value.

For example, 'redistribute static'configuration doesn't specify a route-map
value but OVSDB expects it to have UUID as the value even though Route Map table is empty.
Not setting the value to UUID of the Route Map table entry caused CLI
implementaion issues because of which an empty row was created in Route Map table
everytime a redistribute <protocol> CLI was configured without 'route-map <name>'
so that the value can be a UUID. This implementation adds an entry in Route Map table with
no name and UUID points to this row. As Route_Map is a root table, the incorrect entry is 
not deleted from the DB by ovsdb's garbage collector and 'name' column 
is an index and cannot be empty .This will have unintended consequences
for DC (config persistence) and prevent startup config from being applied
or saved.

Fix : Having separate columns for 'redistribute <protocol>' and
'redistribute <protocol> route-map <name>' will fix the issue.
This commit includes the CLI fixes for the schema changes in the review
https://review.openswitch.net/#/c/10067/

Change-Id: I0dcf34240f612ceed4c83fd2b8263a4a3852b6f6
Signed-off-by: vijavaru <varun.vijayakumar@hpe.com>