fix: usr: Fix for the "show vrf" table_id issue when VRF is down and indentation fix when displaying loopback interfaces.
1. In the case of VRF being down the table_id pointer is NULL
and would hit segmentation fault if tried to access the value.
As a fix, if the VRF is not ready the table_id field will be
shown empty representing that the table_id is not assigned.

2. When loopback interfaces with maximum ID is created the
display of the "show vrf" is not proper. This commit includes
the fix for that.

Change-Id: I6914edf3c5df015538701a08f87d6e90938f21a9
Signed-off-by: Jayachandra Babu K <jayachandra.bab.k@hpe.com>